adding more association types

+++ activerecord/lib/active_record/associations/belongs_to_association.rb
@@ -11,6 +11,9 @@ module ActiveRecord
         when :destroy
           target.destroy
           raise ActiveRecord::Rollback unless target.destroyed?
+        when :destroy_async
+         ActiveRecord::DestroyAssociationLaterJob.
+          perform_later([target])
         else
           target.send(options[:dependent])
         end
